requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-http
python-dotenv
requests
apache-airflow-providers-amazon
minio
apache-airflow-providers-docker